+++ packages/common/src/models/Demos.ts
@@ -78,6 +78,8 @@ export interface Step extends IOpenWebsite, IImagePreview, ITerminal {
   sessionId?: string;
   pollId?: string;
   startOnOpen?: boolean;
+  title?: string;
+  type?: EngageTimeMessageType;
 }
 
 export interface ITerminal {
@@ -108,3 +110,5 @@ export interface State {
   key: string;
   value: string;
 }
+
+export type EngageTimeMessageType = 'demo' | 'slide' | 'custom';