+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   PWA remote control view
 - [#287](https://github.com/estruyf/vscode-demo-time/issues/287): Added the `/api/zoom-in` and
   `/api/zoom-out` API endpoints to control the font size in the editor
+- [#288](https://github.com/estruyf/vscode-demo-time/issues/288): Added the `sendEngageTimeMessage`
+  action to send messages to EngageTime timeline
 
 ## [1.10.0] - 2025-10-01